pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.99.9",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.48.0",
]
